Changing way of receveing unchecked items
