+++ webapp/src/app/view/receive/checklist/index.js
@@ -2,7 +2,6 @@ import React, { Component } from 'react'
 import { Row, Col } from 'react-flexbox-grid'
 
 import {List, ListItem} from 'material-ui/List'
-import RaisedButton from 'material-ui/RaisedButton'
 import Check from 'material-ui/svg-icons/action/check-circle'
 import CheckListIcon from 'material-ui/svg-icons/action/info'
 
@@ -17,9 +16,13 @@ export default class CheckList extends Component {
         {title: 'Caixa de pincéis: preto', check: false, icon: <CheckListIcon />},
         {title: 'Caixa de pincéis: vermelho', check: false, icon: <CheckListIcon />},
         {title: 'Caixa de pincéis: verde', check: false, icon: <CheckListIcon />},
-        {title: 'Apagador de quadro brando', check: false, icon: <CheckListIcon />}
+        {title: 'Apagador de quadro branco', check: false, icon: <CheckListIcon />}
       ],
     }
+  }
+
+  componentDidMount() {
+    this.props.handleItems(this.state.items)
   }
 
   handleCheck = (n, title, check) => {
@@ -47,19 +50,6 @@ export default class CheckList extends Component {
     </div>
   }
 
-  receiveItem = () => {
-    const items = this.state.items;
-    let notCheckedItems = [];
-
-    for(var i = 0; i < items.length; i++) {
-      if(!items[i].check) {
-        notCheckedItems.push(items[i].title)
-      }
-
-      this.props.handleReceive(true, notCheckedItems);
-    }
-  }
-
   render() {
     const styles = {
       list: {
@@ -82,18 +72,11 @@ export default class CheckList extends Component {
             { this.listItems() }
           </List>
         </Col>
-        <Col className="col-fluid" md={12} sm={12} xs={12}>
-          <RaisedButton
-            className="btn-receive-car"
-            label="Recebido"
-            secondary={true}
-            onTouchTap={this.receiveItem} />
-        </Col>
       </Row>
     );
   }
 }
 
 CheckList.propTypes ={
-  handleReceive: React.PropTypes.func.isRequired,
+  handleItems: React.PropTypes.func.isRequired,
 };
