+++ webapp/src/app/view/receive/index.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react'
 import { Grid, Row, Col } from 'react-flexbox-grid'
 
+import RaisedButton from 'material-ui/RaisedButton'
 import OpenKey from './open-key'
 import CheckList from './checklist'
 import AlertDialog from './alert-dialog'
@@ -14,11 +15,25 @@ export default class Receive extends Component {
     this.state = {
       open: false,
       notCheckedItems: [],
+      items: [],
     }
   }
 
-  handleReceive = (open, items) => {
-    this.setState({ open, notCheckedItems: items })
+  handleItems = (items) => {
+    this.setState({ items })
+  }
+
+  handleReceive = () => {
+    const items = this.state.items;
+    let notCheckedItems = [];
+
+    for(var i = 0; i < items.length; i++) {
+      if(!items[i].check) {
+        notCheckedItems.push(items[i].title)
+      }
+    }
+
+    this.setState({ open: true, notCheckedItems })
   }
 
   handleClose = () => {
@@ -33,7 +48,14 @@ export default class Receive extends Component {
             <OpenKey />
           </Col>
           <Col className="col-fluid" md={6} sm={12} xs={12}>
-            <CheckList handleReceive={this.handleReceive}/>
+            <CheckList handleItems={this.handleItems}/>
+          </Col>
+          <Col className="col-fluid" md={12} sm={12} xs={12}>
+            <RaisedButton
+              className="btn-receive-car"
+              label="Recebido"
+              secondary={true}
+              onTouchTap={this.handleReceive} />
           </Col>
         </Row>
         <AlertDialog
